Adds flag to skip lg MQs in IE to breakpoints
Default behavior is true, but we need a way to override this for the grid partial so that it can create and extend the large breakpoint grid classes.

So basically, no one needs this except Responsive, it doesn't affect how anyone calls the breakpoint, and... no one but me probably cares this exists.