Add babelify for es6 style modules & import/export
Babelify is a Browserify transform we can apply to our code during the Browserify task.

Essentially what this does is allow us to leverage all of the es6 features we want via babel presets. So far we have installed the popular `@babel/preset-env` preset which gives us most es6 features. Furthermore, it is a smart preset that has `browserslist` integration where it will look for the browserslist key in your package.json file or a root browserslist file, and match that against es6 features and only supply the features needed for your browser support.

More reading on @babel/preset-env can be found on its website https://babeljs.io/docs/en/next/babel-preset-env.html

This commit adds the babelify funtionality which depends on @babel/core package, and is configured by our .babelrc file which specifies which preset we want to apply. This commit also adds the browserslist key to our package.json file so babel and other front-end tooling such as autoprefixer (not yet in Foundation but might be coming soon) can take advantage of the same list of browser requirements.

The .babelrc file has been separated outside of Gruntfile so that we can migrate off Grunt in the future with little hassle and also provides a way to easily change what presets and options we want to apply without digging through the Gruntfile. This is a standard way of managing babel.

Additional reading: https://medium.com/@hey.aaron/getting-import-export-working-es6-style-using-browserify-babelify-gulp-5hrs-of-life-eca7786e44cc