Add browserify-shim for uncompatible libraries
The browserify-shim package makes CommonJS incompatible files browserifyable (files that don’t support `require` from the CommonJS module syntax). This includes things like our version of jQuery.

The browserify-shim package (paired with the Babelify transform in Browserify) allows us to import the `$` sign from jQuery as an es6 module. Shims are usually specified in package.json as demonstrated on the GitHub page https://github.com/thlorenz/browserify-shim .

For what we do with jQuery, we are depending on it already being enqueued with WordPress or already on the document. So in this case we tell package.json to allow us to use jQuery as a module but import it using the global version attached to the window object, since we’re not including it directly from node_modules.

The browserify-shim package is great for things like jQuery plugins that we want to import into our modules. It is common in the WordPress ecosystem that jQuery is used a lot and there are several slider, modal, and other style plugins that we’d like to use with the es6 syntax. This package allows us to import those packages into our modules and then browserify will bundle them all together in order of the dependencies each module needs.