Revision: Remove unminified compiled version
This commit overwrites the compiled (unminified) version of scripts with the uglified (minified) version.

It doesn’t make sense to keep around both version (minified and unminified) since the unminified is hard to read due to browserify/babelify compiling the original source files.

Instead, we will overwrite the unminified resulting file with the uglified (minified) version.

Also — this commit updates the watch task so that changes in js-dev will re-run the js task.