Add grunt-sass-lint task
This task can be run on its own. We can customize the linting rules as much as we want in the .sasslintrc file. Here is documentation on how to setup rules https://github.com/sasstools/sass-lint/tree/master#rules and here are all the rules https://github.com/sasstools/sass-lint/tree/master/docs/rules (click each MD file to see an explanation).

This task is separate and will need to be run on its own by running `grunt sasslint`. This is because any warning or error will halt Grunt with a message like:

```
Warning:  Use --force to continue.

Aborted due to warnings.
```

Due to this, its best to run this on its own for now.