Update tools-cgb with new icons implementation
We only need to import `icons/package` now. The package file will loop through a sass map of icons, rather than us having to import each group of icons. The map is defined in icons/supported.

To change the icons that a theme supports, overwrite the sass map with the new icons that should exist, then the loop in side `package` will output the correct CSS rules.

Fixes issue with `npm start` in bu-blocks.