Start tracking compiled CSS files
This will help monitor the total file-size. In the next commit, autoprefixer will be added, and it will be useful to see if the package is taking any effect.

Note: This may be removed if tracking compiled CSS files becomes too noisy. Its more for testing purposes at this moment in time.