:bug: Update linear-gradient to new syntax
Updates the generic linear-gradient rule to use the new `to {x}` syntax. The default direction for linear-gradients is `to bottom` so that is what I used in this case.

This removes the autoprefixer warnings we see in Framework to update our linear-gradient rules to adhere to the new standard.

In theory, this mixin doesn’t even need to be used as autoprefixer handles all of these vendor prefixes for us :)